k6runner: split local and http runners to different files
